+++ src/main/java/com/petcare/backend/proyectoIntegrador/entity/Pedido.java
@@ -18,7 +18,7 @@ public class Pedido {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_pedido")
-    private short idPedido;
+    private Integer idPedido;
     
     @Column(name = "fecha")
     private LocalDateTime fecha;
